Fix nightly daily sync SQL cast and harden IAP purchase flow.
Migration 028 fixes ensure_daily_puzzle SMALLINT overload matching so the UTC cron job can generate puzzles; CI now applies pending migrations and the pipeline rolls back before marking rollout failed.

+++ supabase/migrations/026_daily_puzzle_resilience.sql
@@ -224,7 +224,7 @@ BEGIN
   END LOOP;
 
   BEGIN
-    RETURN public.generate_daily_puzzle_fast(p_date, 3, 3, 200);
+    RETURN public.generate_daily_puzzle_fast(p_date, 3::SMALLINT, 3, 200);
   EXCEPTION
     WHEN unique_violation THEN
       SELECT id INTO v_id
@@ -240,7 +240,7 @@ BEGIN
   END;
 
   BEGIN
-    RETURN public.generate_daily_puzzle_fast(p_date, 3, 1, 300);
+    RETURN public.generate_daily_puzzle_fast(p_date, 3::SMALLINT, 1, 300);
   EXCEPTION
     WHEN unique_violation THEN
       SELECT id INTO v_id
